+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
-* To version: 	0.0.4
+* To version: 	0.0.5
 * From version:	0.0.3
+
+# 0.0.5
+* Small fix in getLineNumbers()
+* Added borders etc.
+* You can now walk back with: alt-q
+* File open -dialog width is increased
+* Close open dialogs when quitting, tk_chooseColor is problematic though
+	(it might have own temporary root)
+* Dumped 'proxy-undo',  reason being lag-increase, files are in /dev if interested
+* If get motivation to continue with undo, I will go with ungrid, also in /dev
+
 
 # 0.0.4
 * Made decision not to try to do tab-specific undo, for reasons you can read in dev/todo.py